:fire: Refactored Game of Life Logic into useAutomata
