+++ src/components/Grid.tsx
@@ -1,19 +1,29 @@
 import React from "react";
 import { Box } from "@chakra-ui/core";
+import Cell from "./Cell";
 
 interface Props {
-  col: number;
   cellSize: number;
+  cells: number[][];
+  updateGrid: (i: number, j: number) => void;
 }
 
-const Grid: React.FC<Props> = ({ children, col, cellSize }) => {
+const Grid: React.FC<Props> = ({ cells, cellSize, updateGrid }) => {
   return (
     <Box
       display="grid"
-      gridTemplateColumns={`repeat(${col}, ${cellSize}px)`}
+      gridTemplateColumns={`repeat(${cells[0].length}, ${cellSize}px)`}
       marginBottom={12}
     >
-      {children}
+      {cells.map((rows, i) =>
+        rows.map((_, j) => (
+          <Cell
+            isAlive={!!cells[i][j]}
+            updateGrid={() => updateGrid(i, j)}
+            cellSize={cellSize}
+          />
+        ))
+      )}
     </Box>
   );
 };
